+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changes
 4.10 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Improve test coverage show deprecation warnings when importing modules which
+  are not ported to Python 3.
 
 
 4.9 (2021-01-08)